CLN: assorted
