+++ pandas/core/groupby/ops.py
@@ -207,6 +207,7 @@ class WrappedCythonOp:
             if how in ["var", "mean"] or (
                 self.kind == "transform" and self.has_dropped_na
             ):
+                # has_dropped_na check need for test_null_group_str_transformer
                 # result may still include NaN, so we have to cast
                 values = ensure_float64(values)
